[level 2] Title: 기능개발, Time: 3.32 ms, Memory: 87.1 MB -BaekjoonHub

+++ 프로그래머스/2/42586. 기능개발/기능개발.java
@@ -4,8 +4,8 @@ class Solution {
     public int[] solution(int[] progresses, int[] speeds) {
         Queue<Integer> finishDays = new LinkedList<>();
         for(int i=0; i<speeds.length; i++) {
-            finishDays.add((int)Math.ceil((100 - progresses[i]) / (double) (speeds[i])));
-            System.out.println((int)Math.ceil((100 - progresses[i]) / speeds[i]));
+            finishDays.add((int)(Math.ceil((100 - progresses[i]) / (double) (speeds[i]))));
+            System.out.println((int)(Math.ceil((100 - progresses[i]) / (double) (speeds[i]))));
         }
         List<Integer> list = new ArrayList<>();
         while(!finishDays.isEmpty()) {
